+++ components/test_raftstore-v2/src/server.rs
@@ -232,7 +232,7 @@ impl<EK: KvEngine> RaftExtension for TestExtension<EK> {
         region_epoch: metapb::RegionEpoch,
         split_keys: Vec<Vec<u8>>,
         source: String,
-        encrypt: u16,
+        encrypt: u32,
     ) -> futures::future::BoxFuture<'static, storage::kv::Result<Vec<metapb::Region>>> {
         self.extension
             .split(region_id, region_epoch, split_keys, source, encrypt)
